+++ schemas/tables.sql
@@ -3,13 +3,14 @@
 DROP TABLE IF EXISTS events;
 DROP TABLE IF EXISTS locations;
 DROP TABLE IF EXISTS phases;
+DROP TABLE IF EXISTS config;
 
 CREATE TABLE locations (
   id SERIAL PRIMARY KEY,
   name TEXT NOT NULL,
   address TEXT NOT NULL,
   url TEXT NOT NULL,
-  type VARCHAR(255) NOT NULL
+  type TEXT NOT NULL
 );
 
 CREATE TABLE events (
@@ -17,7 +18,7 @@ CREATE TABLE events (
   title TEXT NOT NULL,
   event_date TIMESTAMP WITH TIME ZONE NOT NULL,
   description TEXT,
-  type VARCHAR(255) NOT NULL,
+  type TEXT NOT NULL,
   is_public BOOLEAN NOT NULL,
   is_mandatory BOOLEAN NOT NULL,
   location_id INT,
@@ -30,4 +31,9 @@ CREATE TABLE phases (
   description TEXT,
   date_from TIMESTAMP WITH TIME ZONE NOT NULL,
   date_to TIMESTAMP WITH TIME ZONE NOT NULL
+);
+
+CREATE TABLE config (
+  key TEXT PRIMARY KEY,
+  value TEXT NOT NULL
 );
