--- src/lib/server/content/config.toml
@@ -1,1 +1,0 @@
-location = "DUS"
