--- src/lib/server/content/term.toml
@@ -1,4 +1,0 @@
-title = "Summer Term 2024"
-batchNumber = 10
-startDate = 2024-04-27T00:00:00Z
-firstWeek = 0
