+++ src/lib/types/config.ts
@@ -1,0 +1,4 @@
+export class ConfigItem {
+    key!: string;
+    value!: string;
+}
