+++ src/routes/api/term/+server.ts
@@ -1,27 +1,56 @@
-import { json } from '@sveltejs/kit';
+import { error, json } from '@sveltejs/kit';
 import fs from 'fs/promises';
 import { access } from 'fs/promises';
 import path from 'path';
 import toml from 'toml';
-import type { Term } from '$lib/types/term.ts';
+import { Term } from '$lib/types/term';
+import sql from '$lib/db';
+import type { ConfigItem } from '$lib/types/config';
 
 export async function GET(e) {
     try {
-        const termFile = path.resolve("./src/lib/server/content/term.toml");
+        const rawConfig = await sql<ConfigItem[]>`
+            SELECT * FROM config
+            WHERE key LIKE 'term.%'
+        `;
 
-        try {
-            await access(termFile);
-        } catch (error) {
-            return json({ error: "Term configuration not found" }, { status: 404 });
+        if (!rawConfig) {
+            error(404, {
+                message: "Config not found",
+            });
         }
 
-        const content = await fs.readFile(termFile, "utf-8");
-        const term = toml.parse(content) as Term;
+        let term = new Term();
 
+        rawConfig.forEach(({key, value}) => {
+            switch (key) {
+                case "term.title":
+                    term.title = value;
+                    break;
+                case "term.batchNumber":
+                    term.batchNumber = parseInt(value);
+                    break;
+                case "term.startDate":
+                    term.startDate = new Date(value);
+                    break;
+                case "term.firstWeek":
+                    term.firstWeek = parseInt(value);
+                    break;
+            }
+        });
+
+        if (!term.title || term.batchNumber === undefined || !term.startDate || term.firstWeek === undefined) {
+            error(404, {
+                message: "Term not found",
+            });
+        }
+    
         return json(term);
-    } catch (error) {
-        console.error("Error reading term", error);
-        
-        return json({ error: "Internal server error" }, { status: 500 });
+    } catch (e) {
+        console.error("Error reading term config", e);
+
+        error(500, {
+            message: "Internal server error",
+        });
     }
 }
